+++ mods/webapp/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.52.1](https://github.com/fonoster/qcobro/compare/v1.52.0...v1.52.1) (2026-09-18)
+
+**Note:** Version bump only for package @qcobro/webapp
+
+# [1.52.0](https://github.com/fonoster/qcobro/compare/v1.51.0...v1.52.0) (2026-09-18)
+
+### Features
+
+- Voz IA barge-in, es-419 default and multilingual language option ([#190](https://github.com/fonoster/qcobro/issues/190)) ([1f66879](https://github.com/fonoster/qcobro/commit/1f668791ebc18fcafdf832cd7299035e95e79a59))
+
+# [1.50.0](https://github.com/fonoster/qcobro/compare/v1.49.0...v1.50.0) (2026-09-12)
+
+### Features
+
+- **webapp:** collapsible sidebar nav ([#182](https://github.com/fonoster/qcobro/issues/182)) ([4808aff](https://github.com/fonoster/qcobro/commit/4808aff11d0432929d6b7f51d38b765df23a19e2))
+
+# [1.49.0](https://github.com/fonoster/qcobro/compare/v1.48.2...v1.49.0) (2026-09-12)
+
+### Features
+
+- **voice:** activate Fonoster answering-machine detection (AMD) ([#181](https://github.com/fonoster/qcobro/issues/181)) ([cca6c1d](https://github.com/fonoster/qcobro/commit/cca6c1d1ee0524b1a9c2d1a5e8ae3b4fcf3b16ae)), closes [#893](https://github.com/fonoster/qcobro/issues/893) [#893](https://github.com/fonoster/qcobro/issues/893) [#893](https://github.com/fonoster/qcobro/issues/893) [#83](https://github.com/fonoster/qcobro/issues/83) [fonoster/fonoster#893](https://github.com/fonoster/fonoster/issues/893) [#893](https://github.com/fonoster/qcobro/issues/893) [fonoster/fonoster#897](https://github.com/fonoster/fonoster/issues/897)
+
+## [1.48.2](https://github.com/fonoster/qcobro/compare/v1.48.1...v1.48.2) (2026-09-12)
+
+### Bug Fixes
+
+- **webapp:** stop showing {{variables}} on fields nobody renders ([#179](https://github.com/fonoster/qcobro/issues/179)) ([f1531cb](https://github.com/fonoster/qcobro/commit/f1531cb762328e6b6053a35e1a0a5dc511c1fa7c))
+
 ## [1.48.1](https://github.com/fonoster/qcobro/compare/v1.48.0...v1.48.1) (2026-09-11)
 
 ### Bug Fixes